+++ troncli/utils.py
@@ -6,7 +6,7 @@ import sys
 import psutil
 import subprocess
 import re
-from colorama import Fore, Style
+from colorama import Fore, Style, Back
 from tqdm import tqdm
 
 import urllib3
@@ -39,6 +39,10 @@ def logo_shadow():
 
 def progress_msg(content):
     print(Fore.CYAN + '[ TRON-CLI ]: ' + content + '...' + Fore.RESET)
+
+
+def imode_msg(content):
+    print(Back.BLUE + Fore.WHITE + Style.BRIGHT + '[ I-MODE ]: ' + Style.NORMAL + content + Fore.RESET + Back.RESET + Style.RESET_ALL)
 
 
 def success_msg(content):
@@ -152,7 +156,10 @@ class Node(object):
                                          'gridport': 18891,
                                          'dbname': 'Null',
                                          'dbusername': 'Null',
-                                         'dbpassword': 'Null'}}
+                                         'dbpassword': 'Null'},
+                             'init_ed': False,
+                             'config_ed': False
+                             }
 
     def get(self):
         return self.node_list
@@ -177,6 +184,14 @@ class Node(object):
                                     'dbname': 'Null',
                                     'dbusername': 'Null',
                                     'dbpassword': 'Null'}
+        self.save()
+
+    async def update_init_done(self, flag):
+        self.node_list['init_ed'] = flag
+        self.save()
+
+    async def update_config_done(self, flag):
+        self.node_list['config_ed'] = flag
         self.save()
 
     async def update_node_version(self, version):