+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+#### Version 1.0.0
+
+##### highlight:
+
+* Add interactive mode.
+
+* Check and throw error messages with insturction when tron-cli has not initialized or configed.
+
+* Support python version 3.6.
+
+##### details:
+
+[X] Add check init and config in config and run, throw error and instruction if catched.
+
+[X] Add interactive mode for whole walk through with key selections, which should cover most use cases.
+
+[X] Fix private event node config error on local host port.
+
+[X] quick - support pass in nettype
+
+[X] support python version 3.6
+
 #### Version 0.3.1
 
 [X] config - set a default committee proposal list - enable CreationOfContracts, DelegateResource, SameTokenName, TvmTransferTrc10 for private testnet